# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProctorAI - Online proctored exam platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List all exams for the current instructor
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create a new exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "201":
          description: Created exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{examId}:
    get:
      operationId: getExam
      tags: [exams]
      summary: Get exam by ID
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamWithQuestions"
        "404":
          description: Not found
    patch:
      operationId: updateExam
      tags: [exams]
      summary: Update exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamUpdate"
      responses:
        "200":
          description: Updated exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteExam
      tags: [exams]
      summary: Delete exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /exams/{examId}/publish:
    post:
      operationId: publishExam
      tags: [exams]
      summary: Publish exam and generate access codes for students
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishExamInput"
      responses:
        "200":
          description: Published exam with access codes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResult"

  # ── Question CRUD ─────────────────────────────────────────────────────────────
...
  /sessions/{sessionId}/flags:
    get:
      operationId: listSessionFlags
      tags: [flags]
      summary: List flags for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Flags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheatingFlag"
    post:
      operationId: reportFlag
      tags: [flags]
      summary: Report a cheating flag (from AI monitor)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagInput"
      responses:
        "201":
          description: Flag recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheatingFlag"
  /flags/{flagId}/review:
    patch:
      operationId: reviewFlag
      tags: [flags]
      summary: Instructor marks a flag as reviewed (dismissed or confirmed)
      parameters:
        - name: flagId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagReviewInput"
      responses:
        "200":
          description: Updated flag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheatingFlag"

  # ── Dashboard / Analytics ───────────────────────────────────────────────────
...
  /exams/{examId}/results:
    get:
      operationId: getExamResults
      tags: [sessions]
      summary: Get all results and flag summary for an exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResults"

  # ── User Profile ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Exam:
      type: object
      required: [id, title, status, durationMinutes, instructorClerkId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published, archived]
        durationMinutes:
          type: integer
        subject:
          type: ["string", "null"]
        instructorClerkId:
          type: string
        questionCount:
          type: integer
        sessionCount:
          type: integer
        flagCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ExamWithQuestions:
      type: object
      required: [id, title, status, durationMinutes, instructorClerkId, createdAt, questions]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published, archived]
        durationMinutes:
          type: integer
        subject:
          type: ["string", "null"]
        instructorClerkId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
    ExamInput:
      type: object
      required: [title, durationMinutes]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        subject:
          type: string
        durationMinutes:
          type: integer
          minimum: 5
    ExamUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        subject:
          type: string
        durationMinutes:
          type: integer
          minimum: 5
        status:
          type: string
          enum: [draft, published, archived]
    PublishExamInput:
      type: object
      required: [studentEmails]
      properties:
        studentEmails:
          type: array
          items:
            type: string
    PublishResult:
      type: object
      required: [exam, accessCodes]
      properties:
        exam:
          $ref: "#/components/schemas/Exam"
        accessCodes:
          type: array
          items:
            $ref: "#/components/schemas/AccessCode"
    AccessCode:
      type: object
      required: [code, studentEmail]
      properties:
        code:
          type: string
        studentEmail:
          type: string

    # ── Question ──────────────────────────────────────────────────────────────
    Question:
      type: object
      required: [id, examId, type, text, orderIndex]
      properties:
        id:
          type: integer
        examId:
          type: integer
        type:
          type: string
          enum: [multiple_choice, true_false, short_answer, essay]
        text:
          type: string
        options:
          type: ["array", "null"]
          items:
            type: string
        correctAnswer:
          type: ["string", "null"]
        points:
          type: integer
        orderIndex:
          type: integer
...
    ExamSession:
      type: object
      required: [id, examId, studentClerkId, status, createdAt]
      properties:
        id:
          type: integer
        examId:
          type: integer
        studentClerkId:
          type: string
        studentEmail:
          type: ["string", "null"]
        studentName:
          type: ["string", "null"]
        accessCode:
          type: string
        status:
          type: string
          enum: [pending, active, submitted, expired]
        score:
          type: ["number", "null"]
        maxScore:
          type: ["integer", "null"]
        flagCount:
          type: integer
        startedAt:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    CheatingFlag:
      type: object
      required: [id, sessionId, type, detectedAt, reviewStatus]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        type:
          type: string
          enum: [face_not_visible, multiple_faces, looking_away, phone_detected, suspicious_movement, audio_anomaly]
        description:
          type: ["string", "null"]
        clipData:
          type: ["string", "null"]
        detectedAt:
          type: string
        reviewStatus:
          type: string
          enum: [pending, dismissed, confirmed]
        reviewedAt:
          type: ["string", "null"]
        reviewNote:
          type: ["string", "null"]
    FlagInput:
      type: object
      required: [type, detectedAt]
      properties:
        type:
          type: string
          enum: [face_not_visible, multiple_faces, looking_away, phone_detected, suspicious_movement, audio_anomaly]
        description:
          type: string
        clipData:
          type: string
        detectedAt:
          type: string
    FlagReviewInput:
      type: object
      required: [reviewStatus]
      properties:
        reviewStatus:
          type: string
          enum: [dismissed, confirmed]
        reviewNote:
          type: string

    # ── Dashboard ─────────────────────────────────────────────────────────────
...
    ExamResults:
      type: object
      required: [exam, sessions, totalFlags, avgScore, submittedCount]
      properties:
        exam:
          $ref: "#/components/schemas/Exam"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/ExamSession"
        totalFlags:
          type: integer
        avgScore:
          type: number
        submittedCount:
          type: integer