# lib/api-zod/src/generated/types/accessCode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ProctorAI - Online proctored exam platform
 * OpenAPI spec version: 0.1.0
 */

export interface AccessCode {
  code: string;
  studentEmail: string;
}