# lib/api-zod/src/generated/types/examUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ProctorAI - Online proctored exam platform
 * OpenAPI spec version: 0.1.0
 */
import type { ExamUpdateStatus } from './examUpdateStatus';

export interface ExamUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  subject?: string;
  /** @minimum 5 */
  durationMinutes?: number;
  status?: ExamUpdateStatus;
}