# lib/api-zod/src/generated/types/flagReviewInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ProctorAI - Online proctored exam platform
 * OpenAPI spec version: 0.1.0
 */
import type { FlagReviewInputReviewStatus } from './flagReviewInputReviewStatus';

export interface FlagReviewInput {
  reviewStatus: FlagReviewInputReviewStatus;
  reviewNote?: string;
}