# lib/api-zod/src/generated/types/publishResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ProctorAI - Online proctored exam platform
 * OpenAPI spec version: 0.1.0
 */
import type { AccessCode } from './accessCode';
import type { Exam } from './exam';

export interface PublishResult {
  exam: Exam;
  accessCodes: AccessCode[];
}